Moved all settings objects to separate files.
