+++ mirp/dicomImport.py
@@ -11,7 +11,7 @@ from mirp.contourClass import ContourClassDeprecated
 from mirp.imageSUV import SUVscalingObj
 from mirp.imageMetaData import get_pydicom_meta_tag, has_pydicom_meta_tag
 from mirp.imageClass import ImageClass
-from mirp.settings.settingsClass import SettingsClass
+from mirp.settings.settingsGeneric import SettingsClass
 from mirp.roiClass import RoiClass
 from mirp.utilities import parse_roi_name
 from pydicom import FileDataset
