+++ mirp/featureSets/intensityVolumeHistogram.py
@@ -5,7 +5,7 @@ from typing import Tuple, Optional
 
 from mirp.imageClass import ImageClass
 from mirp.roiClass import RoiClass
-from mirp.settings.settingsClass import FeatureExtractionSettingsClass
+from mirp.settings.settingsFeatureExtraction import FeatureExtractionSettingsClass
 from mirp.images.genericImage import GenericImage
 from mirp.images.transformedImage import TransformedImage
 from mirp.masks.baseMask import BaseMask
