+++ mirp/featureSets/volumeMorphology.py
@@ -3,7 +3,7 @@ import pandas as pd
 
 from mirp.imageClass import ImageClass
 from mirp.roiClass import RoiClass
-from mirp.settings.settingsClass import FeatureExtractionSettingsClass
+from mirp.settings.settingsFeatureExtraction import FeatureExtractionSettingsClass
 from mirp.images.genericImage import GenericImage
 from mirp.masks.baseMask import BaseMask
