+++ mirp/imageProcess.py
@@ -2,7 +2,8 @@ import warnings
 from typing import Union, List, Tuple, Optional, Any
 from copy import deepcopy
 
-from mirp.settings.settingsClass import SettingsClass, FeatureExtractionSettingsClass
+from mirp.settings.settingsGeneric import SettingsClass
+from mirp.settings.settingsFeatureExtraction import FeatureExtractionSettingsClass
 from mirp.imageClass import ImageClass
 from mirp.roiClass import RoiClass
